+++ src/Cool/View/View.cpp
@@ -117,8 +117,8 @@ auto View::contains(ViewCoordinates pos, ImageSizeInsideView image_size) -> bool
     const auto img_size = image_size.fit_into(*_size);
 
     const auto pos_in_img = pos + glm::vec2{
-                                (img_size.width() - _size->width()) * 0.5f,
-                                (img_size.height() - _size->height()) * 0.5f,
+                                (img_size.width() - static_cast<float>(_size->width())) * 0.5f,
+                                (img_size.height() - static_cast<float>(_size->height())) * 0.5f,
                             };
     return pos_in_img.x >= 0.f && pos_in_img.x <= img_size.width()
            && pos_in_img.y >= 0.f && pos_in_img.y <= img_size.height();
